requirements.txt
requests>=2.25.0
mysql-connector-python>=8.0.0
pinatapy-unofficial>=0.1.0
click>=8.0.0
colorama>=0.4.4
google-generativeai>=0.5.0
pytest>=6.0.0